Make Domain Android-portable: drop CGFloat and LocalizedError
Part A: Replace all CGFloat in ReaderPreferences with Double, removing
the #if canImport(CoreGraphics) guard and its Android typealias workaround.
Adjust LayoutSettingsModel and VisualInspectorScreen call-sites to use
CGFloat(...) where the type boundary requires an explicit conversion.

Part B: Remove the #if canImport(CoreGraphics) LocalizedError extension
from DomainError entirely. Move error message strings into the UI layer:
LibraryErrorPresentation gets explicit cases for the remaining DomainError
variants; ReaderViewModel.errorMessage adds the missing default cases.
DomainErrorTests updated to drop the now-meaningless localizedDescription
assertion.

Part C: LibraryStore security-scoped resource guard changed from
#if canImport(CoreGraphics) to #if !os(Android), which is semantically
clearer about the Android-specific exclusion.

Verified: 90/90 Library tests pass, Android both .so files rebuild
cleanly, iOS app BUILD SUCCEEDED.

+++ Packages/Features/Reader/Sources/Reader/LayoutSettingsModel.swift
@@ -32,14 +32,14 @@ final class LayoutSettingsModel {
     }
 
     func incrementStaffSize() async {
-        let next = min(staffSize + 1, ReaderPreferences.maxStaffSize)
+        let next = min(staffSize + 1, CGFloat(ReaderPreferences.maxStaffSize))
         guard next != staffSize else { return }
         staffSize = next
         await onChange?()
     }
 
     func decrementStaffSize() async {
-        let next = max(staffSize - 1, ReaderPreferences.minStaffSize)
+        let next = max(staffSize - 1, CGFloat(ReaderPreferences.minStaffSize))
         guard next != staffSize else { return }
         staffSize = next
         await onChange?()
