+++ Packages/Features/Reader/Sources/Reader/Screens/VisualInspectorScreen.swift
@@ -101,7 +101,7 @@ struct VisualInspectorScreen: View {
         )
         Stepper(
             value: staffSize,
-            in: ReaderPreferences.minStaffSize ... ReaderPreferences.maxStaffSize,
+            in: CGFloat(ReaderPreferences.minStaffSize) ... CGFloat(ReaderPreferences.maxStaffSize),
             step: 1,
         ) {
             Text(String(
